+++ sr_port/ionl_wteol.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2004 Sanchez Computer Associates, Inc.	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -16,7 +16,7 @@
 #include "iottdef.h"
 
 /* essentially the same as iott_wteol */
-void ionl_wteol(short val, io_desc *io_ptr)
+void ionl_wteol(int4 val, io_desc *io_ptr)
 {
 	mstr	eol;
 	int	eol_cnt;
