+++ sr_unix/iott_write.c
@@ -1,6 +1,6 @@
 /****************************************************************
  *								*
- *	Copyright 2001 Sanchez Computer Associates, Inc.	*
+ *	Copyright 2001, 2004 Sanchez Computer Associates, Inc.	*
  *								*
  *	This source code contains the intellectual property	*
  *	of its copyright holder(s), and is made available	*
@@ -119,7 +119,7 @@ void iott_write(mstr *v)
 					io_ptr->dollar.y++;
 					io_ptr->dollar.x = 0;
 				}
-				if ((START != io_ptr->esc_state) || ((int)(io_ptr->dollar.x + str_len) <= (int)io_ptr->width))
+				if ((START != io_ptr->esc_state) || ((io_ptr->dollar.x + str_len) <= io_ptr->width))
 					len = str_len;
 				else
 					len = io_ptr->width - io_ptr->dollar.x;
